requirements.txt
asgiref==3.8.1
boto3==1.35.6
botocore==1.35.6
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
dj-database-url==2.2.0
Django==4.2.11
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-currentuser==0.8.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-protected-media==1.0.2
django-rest-passwordreset==1.4.1
django-storages==1.14.4
django-taggit==6.0.0
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
getmac==0.9.5
gunicorn==23.0.0
idna==3.8
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
ua-parser==0.18.0
Unidecode==1.3.8
urllib3==1.26.19
user-agents==2.2.0
whitenoise==6.7.0
